refact: Propose new architecture for backend
Aims to simplify & avoid useless repetitions
Keep endpoint clean of any technical stuff
