--- backend/src/__init__.py
@@ -1,3 +1,0 @@
-from fastapi import APIRouter
-
-router = APIRouter()
