--- backend/src/adapters/stores/models/__init__.py
@@ -1,5 +1,0 @@
-from sqlalchemy.orm import DeclarativeBase
-
-
-class Base(DeclarativeBase):
-    pass
