+++ backend/src/core/stores.py
@@ -1,6 +1,8 @@
-from typing import Any, Iterable
+from collections.abc import Iterable
+from typing import Any
 
 from pydantic import BaseModel
+from sqlalchemy.orm import DeclarativeBase
 
 
 class StoreOrderClause(BaseModel):
@@ -63,3 +65,7 @@ def StoreIn(key: str, value: Iterable[Any], **kwargs) -> StoreFilter:
         del kwargs["op"]
 
     return StoreFilter(key=key, value=value, op="in", **kwargs)
+
+
+class Base(DeclarativeBase):
+    pass
