--- backend/src/domain/entities/__init__.py
@@ -1,43 +1,0 @@
-from .events import (
-    Event,
-    Kind as EventKind,
-)
-from .flows import Flow, Step
-from .labels import Label, LabelScrapping, LabelSrcKind
-from .metrics import (
-    Kind as MetricKind,
-    Metric,
-)
-from .notes import Block, Note
-from .stores import (
-    StoreEqual,
-    StoreFilter,
-    StoreGreater,
-    StoreIn,
-    StoreLower,
-    StoreNotEqual,
-    StoreOrder,
-    StoreOrderClause,
-)
-
-__all__ = [
-    "Flow",
-    "Step",
-    "StoreOrderClause",
-    "Label",
-    "LabelSrcKind",
-    "StoreIn",
-    "StoreFilter",
-    "StoreEqual",
-    "StoreGreater",
-    "StoreLower",
-    "StoreNotEqual",
-    "StoreOrder",
-    "LabelScrapping",
-    "Metric",
-    "MetricKind",
-    "Event",
-    "EventKind",
-    "Note",
-    "Block",
-]
