--- backend/src/domain/usecases/__init__.py
@@ -1,7 +1,0 @@
-from .events import Events
-from .labels import Labels
-from .metrics import Metrics
-from .notes import Notes
-from .workflows import Flows, Steps
-
-__all__ = ["Labels", "Metrics", "Flows", "Steps", "Events", "Notes"]
