+++ backend/src/events/__init__.py
@@ -1,0 +1,4 @@
+from .event import Events
+from .entities import Event, EventKind
+
+__all__ = ["Events", "Event", "EventKind"]
