+++ backend/src/events/entities.py
@@ -4,7 +4,7 @@ from enum import Enum
 from pydantic import BaseModel
 
 
-class Kind(str, Enum):
+class EventKind(str, Enum):
     incident = "incident"
     maintenance = "maintenance"
 
@@ -13,7 +13,7 @@ class Event(BaseModel):
     id: int
     title: str
     description: str | None = None
-    kind: Kind
+    kind: EventKind
     labels: dict[str, str] = {}
     started_at: datetime
     ended_at: datetime
