+++ backend/src/events/model.py
@@ -3,9 +3,12 @@ from datetime import UTC, datetime
 from sqlalchemy import JSON, DateTime, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column
 
-from src.domain import entities
+from core.stores import Base
 
-from . import Base
+from .entities import (
+    Event as EventEntity,
+    EventKind,
+)
 
 
 class Event(Base):
@@ -16,16 +19,18 @@ class Event(Base):
     description: Mapped[str | None] = mapped_column(String)
     kind: Mapped[str] = mapped_column(String, nullable=False)
     labels: Mapped[dict[str, str]] = mapped_column(JSON, nullable=False, default={})
-    created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False, default=lambda: datetime.now(UTC))
+    created_at: Mapped[datetime] = mapped_column(
+        DateTime(timezone=True), nullable=False, default=lambda: datetime.now(UTC)
+    )
     started_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False)
     ended_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False)
 
-    def to_entity(self) -> entities.Event:
-        return entities.Event(
+    def to_entity(self) -> EventEntity:
+        return EventEntity(
             id=self.id,
             title=self.title,
             description=self.description,
-            kind=entities.EventKind(self.kind),
+            kind=EventKind(self.kind),
             labels=self.labels,
             started_at=self.started_at,
             ended_at=self.ended_at,
