+++ backend/src/facets/__init__.py
@@ -1,0 +1,4 @@
+from .facets import Facets
+from .entities import Metric, MetricKind, Label, LabelSrcKind, LabelScrapping
+
+__all__ = ["Facets", "Metric", "MetricKind", "Label", "LabelSrcKind", "LabelScrapping"]
