+++ backend/src/facets/entities.py
@@ -4,6 +4,25 @@ from typing import TypedDict
 from pydantic import BaseModel, Field
 
 
+class MetricKind(str, Enum):
+    counter = "counter"
+    measure = "measure"
+    stats = "stats"
+    other = "other"
+
+
+class Metric(BaseModel):
+    id: int
+    key: str
+    kind: MetricKind
+    scope: str
+    unit: str
+    display_as: str
+    description: str
+    associated_identifiers: list[str] = []
+    labels: list[int] = []
+
+
 class LabelSrcKind(str, Enum):
     workflow = "workflow"
     metrics = "metrics"
