+++ backend/src/notebooks/__init__.py
@@ -1,0 +1,5 @@
+from .notebook import Notebooks
+from .exceptions import NoteLockedException
+from .entities import Block, Note
+
+__all__ = ["Notebooks", "NoteLockedException", "Block", "Note"]
